NeuroEase/requirements.txt
Flask==2.0.2
Flask-SocketIO==5.1.1
tensorflow==2.6.0
numpy==1.19.5
opencv-python==4.5.3.56
SpeechRecognition==3.8.1
pyttsx3==2.90
pyautogui==0.9.52
pyserial==3.5
paho-mqtt==1.5.1
sqlite3
pyttsx3==2.90
mediaPipe==0.8.5
python-dotenv==0.19.2
spaCy==3.0.6
